remove ExLimiter.Utils module
