+++ lib/ex_limiter/bucket.ex
@@ -1,7 +1,5 @@
 defmodule ExLimiter.Bucket do
   @moduledoc false
-  alias ExLimiter.Utils
-
   @type t :: %__MODULE__{}
 
   defstruct key: nil,
@@ -9,7 +7,7 @@ defmodule ExLimiter.Bucket do
             last: nil,
             version: %{}
 
-  def new(key), do: %__MODULE__{key: key, last: Utils.now()}
+  def new(key), do: %__MODULE__{key: key, last: System.system_time(:millisecond)}
 
   def new(contents, key) when is_map(contents) do
     struct(__MODULE__, Map.put(contents, :key, key))
